docs: clarify poller's functioning
Full rewording of the main description for the package. Explain the
meaning of polling and how each component of the package behaves,
specifying that Pusher errors are not interruptions for other Pusher
functions passed.

The Poll method was named in the documentation as Poller and this is
corrected with rewording of the whole comment.